+++ android/app/src/tv/kotlin/gitlab/openlyst/doudou/MainActivity.kt
@@ -1,5 +1,11 @@
 package gitlab.openlyst.doudou
 
 import com.ryanheise.audioservice.AudioServiceActivity
+import io.flutter.embedding.engine.FlutterEngine
 
-class MainActivity : AudioServiceActivity()
+class MainActivity : AudioServiceActivity() {
+    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
+        super.configureFlutterEngine(flutterEngine)
+        AudioRouter.register(flutterEngine.dartExecutor.binaryMessenger, applicationContext)
+    }
+}